Migrate automation/tests to Python 3 (#2582)

+++ tests/pytests/test_stats.py
@@ -33,14 +33,14 @@ def runTestWithSeed(env, s=None):
 
     env.expect('FLUSHALL')
     if s == None:
-        s = time()
+        s = int(time())
     env.debugPrint('seed: %s' % str(s), force=True)
     seed(s)
 
     idx = 'idx'
     count = 10000
     cleaning_loops = 5
-    loop_count = count / cleaning_loops
+    loop_count = int(count / cleaning_loops)
 
     ### test increasing integers
     env.expect('ft.config set FORK_GC_CLEAN_THRESHOLD 0').ok()
@@ -53,7 +53,7 @@ def runTestWithSeed(env, s=None):
 
     env.expect('FT.SEARCH idx * LIMIT 0 0').equal([count])
     for i in range(count):
-        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([1L, 'doc%d' % i, ['n', str(i)]])
+        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([1, 'doc%d' % i, ['n', str(i)]])
     #check_not_empty(env, idx)
 
     for i in range(cleaning_loops):
@@ -64,7 +64,7 @@ def runTestWithSeed(env, s=None):
             forceInvokeGC(env, 'idx')
 
     for i in range(count):
-        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([0L])
+        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([0])
     check_empty(env, idx)
 
     ### test random integers
@@ -85,7 +85,7 @@ def runTestWithSeed(env, s=None):
     check_empty(env, idx)
 
     for i in range(count):
-        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([0L])
+        env.expect('FT.SEARCH', 'idx', '@n:[%d,%d]' % (i, i))#.equal([0])
     check_empty(env, idx)
 
     ## test random floats
